+++ melbourneStore/src/main/java/com/melbournestore/adaptors/OrderListAdapter.java
@@ -172,7 +172,7 @@ public class OrderListAdapter extends BaseAdapter {
     }
 
     private void setComponentsStatus(Button plusButton, Button minusButton, int position) {
-        int stock_num = mItems.get(position).getStock();
+        long stock_num = mItems.get(position).getStock();
         int plate_num = mItems.get(position).getUnit();
 
         if (plate_num >= stock_num) {
